Remove client result set meta
